refactor to build tags
