+++ internal/tokenizer/common.go
@@ -1,0 +1,6 @@
+// Package tokenizer implements file tokenization used by the enry content
+// classifier. This package is an implementation detail of enry and should not
+// be imported by other packages.
+package tokenizer
+
+const byteLimit = 100000
