+++ internal/tokenizer/flex/tokenize_c_test.go
@@ -1,0 +1,25 @@
+package flex
+
+// import (
+// 	"testing"
+
+// 	"gopkg.in/src-d/enry.v1/internal/tokenizer"
+// )
+
+// func BenchmarkTokenizerC(b *testing.B) {
+// 	b.ReportAllocs()
+// 	for i := 0; i < b.N; i++ {
+// 		for _, test := range tokenizer.Tests {
+// 			TokenizeC(test.content)
+// 		}
+// 	}
+// }
+
+// func BenchmarkTokenizerFlex(b *testing.B) {
+// 	b.ReportAllocs()
+// 	for i := 0; i < b.N; i++ {
+// 		for _, test := range tokenizer.Tests {
+// 			TokenizeFlex(test.content)
+// 		}
+// 	}
+// }
